+++ packages/papupata/src/main/common-types/jsonAPI.ts
@@ -1,4 +1,5 @@
 import type { JSONApiType } from './jsonAPIType'
+import { ZodTypeAny } from 'zod'
 
 type StringParameterTypes = 'string' | 'boolean' | 'number' | 'date'
 
@@ -29,8 +30,8 @@ export interface JSONAPI {
   method: string
   query: QueryParameter[]
   pathParams: PathParameter[]
-  body?: JSONApiType | null
-  response?: JSONApiType | null
+  body?: JSONApiType | ZodTypeAny | null
+  response?: JSONApiType | ZodTypeAny | null
   alternativeResponses: Array<{
     code: number
     description?: string