+++ packages/papudoc/RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+0.5.0: Support for zod schemas from papupata
+
 0.3.3: Support for string and number constant enum literals (such as the type of a variable being `MyEnum.Value`)
 
 0.3.2-rc1: Inlined internal dependencies